HL-146: add DUMMY_COMPANY_FORM to change created company type during development. Fix renamed fixture in test_user_model.
