+++ .env.benefit.example
@@ -4,6 +4,10 @@ APPLY_MIGRATIONS=1
 CREATE_SUPERUSER=1
 CORS_ORIGIN_ALLOW_ALL=1
 
+# if changing, §, also set MOCK_FLAG=1
+DUMMY_COMPANY_FORM=OY
+MOCK_FLAG=1
+
 ## Authentication
 DISABLE_AUTHENTICATION=1
 OIDC_RP_CLIENT_ID=
