+++ backend/benefit/README.md
@@ -35,6 +35,11 @@ Allow user to create test database
 
 The project is now running at [localhost:8000](http://localhost:8000)
 
+### Testing and debugging
+
+To run the backend without integrations, set MOCK_FLAG=1 in the .env.benefit-backend
+If MOCK_FLAG is set, additionally DUMMY_COMPANY_FORM can be set to test with different company_form parameters.
+
 ## Keeping Python requirements up to date
 
 1. Install `pip-tools`:
