+++ backend/benefit/companies/tests/data/company_data.py
@@ -1,3 +1,13 @@
+from django.conf import settings
+
+
+def get_dummy_company_data():
+    dummy_data = DUMMY_COMPANY_DATA.copy()
+    if hasattr(settings, "DUMMY_COMPANY_FORM"):
+        dummy_data["company_form"] = settings.DUMMY_COMPANY_FORM
+    return dummy_data
+
+
 DUMMY_COMPANY_DATA = {
     "id": "8c0c7a56-cb98-4c31-87ca-6f1853253986",
     "name": "I. Haanpää Oy",
