+++ backend/benefit/helsinkibenefit/settings.py
@@ -80,6 +80,7 @@ env = environ.Env(
     ENABLE_SEND_AUDIT_LOG=(bool, False),
     WKHTMLTOPDF_BIN=(str, "/usr/bin/wkhtmltopdf"),
     DISABLE_AUTHENTICATION=(bool, False),
+    DUMMY_COMPANY_FORM=(str, "OY"),
 )
 if os.path.exists(env_file):
     env.read_env(env_file)
@@ -231,6 +232,7 @@ YTJ_TIMEOUT = env.int("YTJ_TIMEOUT")
 
 # Mock flag for testing purposes
 MOCK_FLAG = env.bool("MOCK_FLAG")
+DUMMY_COMPANY_FORM = env.str("DUMMY_COMPANY_FORM")
 
 # Authentication settings begin
 SESSION_COOKIE_AGE = env.int("SESSION_COOKIE_AGE")
