+++ backend/benefit/users/tests/test_user_model.py
@@ -3,5 +3,5 @@ from django.contrib.auth import get_user_model
 User = get_user_model()
 
 
-def test_user_model(user):
+def test_user_model(bf_user):
     assert User.objects.count() == 1
